requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.0
pandas==2.0.3
pyarrow==12.0.1
numpy==1.24.3
openai==1.12.0
python-dotenv==1.0.0
azure-identity==1.19.0